% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_vector_to_matrix.R
\name{beta_vector_to_matrix}
\alias{beta_vector_to_matrix}
\title{Create \eqn{\beta} matrix from vector of \eqn{\beta} for \eqn{\beta_k : k \neq j}}
\usage{
beta_vector_to_matrix(values, p, J, null_j, beta_j_null = NULL)
}
\arguments{
\item{values}{A vector containing the values for all \eqn{\beta_k}, with \eqn{k \neq j}, as well as all \eqn{\beta_{k0}, for k = 1, \dots, J}.
In particular, this vector should be so that the first \eqn{(J-2)(p+1)} entries are \eqn{\beta_{10}, \beta_{1}^{\top}, \beta_{20}, \beta_{2}^{\top}, \dots, \beta_{(j-1)0}, \beta_{j-1}^{\top},  \beta_{(j+1)0}, \beta_{j+1}^{\top}, \dots,  \beta_{(J-1)0}, \beta_{J-1}^{\top}, \beta_{j0}}.
Then, the \eqn{(J-2)(p+1) + 1} entry should be \eqn{\beta_{j0}}.}

\item{p}{This should be the number of covariates.}

\item{J}{This should be the number of categories.}

\item{null_j}{This specifies for which category you have set \eqn{\beta_j = 0}.}

\item{beta_j_null}{This is the null hypothesized value for \eqn{\beta_j}, which is by default set to be \eqn{\beta_j = 0}.}
}
\value{
The full \eqn{(p+1) \times (J-1)} matrix of \eqn{\beta}
}
\description{
Create \eqn{\beta} matrix from vector of \eqn{\beta} for \eqn{\beta_k : k \neq j}
}
\author{
Shirley Mathur
}
