% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_independent_p_values.R
\name{combine_independent_p_values}
\alias{combine_independent_p_values}
\title{Combine independent p-values that test the same null hypothesis into an overall summary}
\usage{
combine_independent_p_values(ps)
}
\arguments{
\item{ps}{A vector of independent p-values (that test the same null hypothesis)}
}
\value{
A single p-value that combines the evidence from the vector
}
\description{
For a vector of p-values \eqn{(p_1...p_n)} that independently test a specific null hypothesis $H_0$, find the relevant
quantile of a chi-sq distribution for each p-value, then sum them up. That's a \eqn{\chi^2_n}-distributed test statistic under $H_0$,
and the relevant p-value can be found. Useful for combining analyses after stratification (to deal with potentially
complex experimental designs).
}
\author{
Amy Willis
}
