% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcr.R
\docType{data}
\name{ddpcr}
\alias{ddpcr}
\title{Digital droplet PCR measurements and metadata from a study of people living with cystic fibrosis}
\format{
A data frame object with 74 observations of 4 covariates.
\describe{
\item{Value}{droplet digital PCR, giving the observed total bacterial load in the sample in copies/uL}
\item{Treatment}{1 if the sample is from a subject receiving antibiotics, 2 otherwise }
\item{SpecimenType}{the body site of origin of the sample, either saliva or sputum}
\item{Subject}{an ID identifying the participant }
}
}
\usage{
ddpcr
}
\description{
Digital droplet PCR measurements and metadata from a study of people living with cystic fibrosis
}
\references{
Lu, J et al. (2020). \emph{Parallel Analysis of Cystic Fibrosis Sputum and Saliva Reveals Overlapping Communities and an Opportunity for Sample Decontamination}. ASA Journals mSystems, 5, 4. <doi: 10.1128/msystems.00296-20>.
}
\keyword{datasets}
