% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_info_contribution.R
\name{fisher_info_contribution}
\alias{fisher_info_contribution}
\title{Compute fisher information contribution to robust score statistic for glm robust score tests.}
\usage{
fisher_info_contribution(i, model_fits, yy, xx, m = 1, family, link)
}
\arguments{
\item{i}{Index of observation.}

\item{model_fits}{The fitted glm under the null hypothesis.}

\item{yy}{Vector of observed responses.}

\item{xx}{Design matrix for model.}

\item{m}{Number of parameters fixed under null hypothesis.}

\item{family}{The model family for the fitted glm.}

\item{link}{The link function utilized in the fitted glm.}
}
\description{
Compute fisher information contribution to robust score statistic for glm robust score tests.
}
