% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_test.R
\name{gee_test}
\alias{gee_test}
\title{Generalized Estimating Equations under technical replication with robust and non-robust Wald and Rao (score) tests}
\usage{
gee_test(...)
}
\arguments{
\item{...}{Arguments that you would pass to a regular \code{geepack::geeglm} call. Note that for now we only have functionality for Poisson tests with log link}
}
\description{
Generalized Estimating Equations under technical replication with robust and non-robust Wald and Rao (score) tests
}
\examples{
# TODO


}
