% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_statistic.R
\name{get_test_statistic}
\alias{get_test_statistic}
\title{Just in case anyone wants to invert a p-value to recover the chi-squared distributed test statistic}
\usage{
get_test_statistic(pvalue, df)
}
\arguments{
\item{pvalue}{the p-value to invert}

\item{df}{the degrees of freedom for the test}
}
\description{
Just in case anyone wants to invert a p-value to recover the chi-squared distributed test statistic
}
