% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_se.R
\name{jackknife_se}
\alias{jackknife_se}
\title{Jackknife standard errors}
\usage{
jackknife_se(object, dat, id = NULL)
}
\arguments{
\item{object}{The fitted object under the alternative.}

\item{dat}{The data used to fit the model}

\item{id}{Observations with the same id are in the same cluster. If not included, independence between
observations is assumed.}
}
\description{
Jackknife standard errors
}
