% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_presence.R
\docType{data}
\name{light_presence}
\alias{light_presence}
\title{Presence and absence of COG 4251 with function Bacteriophytochrome (light-regulated signal transduction histidine kinase) within a set of \emph{Prochlorococcus} genomes}
\format{
A data frame object with 31 observations of 5 covariates.
\describe{
\item{sample}{Sample id}
\item{cog}{COG, all observations are from COG 4251}
\item{presence}{1 if COG 4251 is present in the sample, 0 otherwise}
\item{clade}{light level, combined with category I or II}
\item{light}{HL for high light condition, LL for low light condition}
}
}
\usage{
light_presence
}
\description{
Presence and absence of COG 4251 with function Bacteriophytochrome (light-regulated signal transduction histidine kinase) within a set of \emph{Prochlorococcus} genomes
}
\references{
https://statdivlab.github.io/blog/articles/ebame10_regression_pt2.html
}
\keyword{datasets}
