% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincom.R
\name{lincom}
\alias{lincom}
\title{Test hypothesis of form A x beta = c}
\usage{
lincom(test_object, A, c)
}
\arguments{
\item{test_object}{Object of type \code{raoFit}.}

\item{A}{Matrix specifying linear combination of parameters for user-specified hypotheses.}

\item{c}{Vector that has user-specified null hypothesis value for user-specified linear combinations of parameters of interest.}
}
\value{
Table with relevant quantities of interest for hypothesis tests.
}
\description{
Test hypothesis of form A x beta = c
}
\author{
Shirley Mathur
}
