% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_get_probs.R
\name{multinom_get_probs}
\alias{multinom_get_probs}
\title{Compute multinomial probabilities for a given value of model parameters.}
\usage{
multinom_get_probs(X, Y, beta)
}
\arguments{
\item{X}{This should be the \eqn{n \times p} design matrix of covariates.}

\item{Y}{This should be the \eqn{n \times J} data matrix of outcomes.}

\item{beta}{This should be the \eqn{(p+1) \times (J-1) \beta} matrix.}
}
\value{
The multinomial probabilities for a given value of \eqn{\beta}.
}
\description{
Compute multinomial probabilities for a given value of model parameters.
}
\author{
Shirley Mathur
}
