% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_info_mat.R
\name{multinom_info_mat}
\alias{multinom_info_mat}
\title{Compute Information matrix for model parameters.}
\usage{
multinom_info_mat(X, Y, probs)
}
\arguments{
\item{X}{This should be the \eqn{n x p} design matrix of covariates.}

\item{Y}{This should be the \eqn{n x J} data matrix of outcomes.}

\item{probs}{This should be the \eqn{n \times J} matrix of estimated probabilities for each sample.}
}
\value{
The Fisher information matrix for model parameters.
}
\description{
Compute Information matrix for model parameters.
}
\author{
Shirley Mathur
}
