% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_mle_strong_null.R
\name{multinom_mle_strong_null}
\alias{multinom_mle_strong_null}
\title{Negative log-likelihood for multinomial data under the strong null (\eqn{\beta_1 = beta_2 = \dots = \beta_{J-1} = 0)}}
\usage{
multinom_mle_strong_null(betanots, Y, X)
}
\arguments{
\item{betanots}{A vector containing the initial values for all \eqn{\beta_{k0}, for k = 1, \dots, J-1}.
In particular, this vector should be so that the entries are ordered as \eqn{\beta_{10}, \beta_{20}, \dots, \beta_{(J-1)0}}.}

\item{Y}{This should be the n x J data matrix of outcomes.}

\item{X}{This should be the n x p design matrix of covariates.}
}
\value{
A vector containing the optimal values for betanots  to maximize the log-likelihood under the null constraint that \eqn{\beta_1 = beta_2 = \dots = \beta_{J-1} = 0}.
The components are listed out in the same manner as in the betanots parameter.
}
\description{
Negative log-likelihood for multinomial data under the strong null (\eqn{\beta_1 = beta_2 = \dots = \beta_{J-1} = 0)}
}
\author{
Shirley Mathur
}
