% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_penalized_estimation.R
\name{multinom_penalized_estimation}
\alias{multinom_penalized_estimation}
\title{Optimization under null or alternative for multinomial model with a Firth penalty.}
\usage{
multinom_penalized_estimation(
  beta,
  X,
  Y,
  null = TRUE,
  strong = FALSE,
  null_j = NULL,
  tol = 1e-05,
  stepSize = 0.5,
  arm_c = 0.5,
  maxit = 250,
  maxit_fs = 5
)
}
\arguments{
\item{beta}{The initial values provided for the \eqn{\beta} parameters.}

\item{X}{The \eqn{n x p} design matrix of covariates.}

\item{Y}{The \eqn{n x J} data matrix of outcomes.}

\item{null}{If TRUE, optimizes under the null, if FALSE, optimizes under the alternative. Defauls to TRUE.}

\item{strong}{If FALSE, this function will compute the robust score statistic to test the weak null that for one specific \eqn{j}, \eqn{\beta_j = 0} for the length \eqn{p} vector \eqn{\beta_j}.
If TRUE, this function instead computes the robust score statistic to test the strong null that \eqn{\beta_1 = \beta_2 = \dots = \beta_{J-1} = 0} for all length \eqn{p} vectors \eqn{\beta_j}, \eqn{j\in\{1,\ldots,J-1\}}.
Default is FALSE.}

\item{null_j}{If \code{strong} is FALSE, this argument must be supplied. This gives the category \eqn{j} in the weak null hypothesis that \eqn{\beta_j = 0}.}

\item{tol}{The tolerance used to determine how much better update function value must be prior to stopping algorithm.}

\item{stepSize}{The size of the step to take during the parameter update step, used in the MLE step.}

\item{arm_c}{Control parameter for checking Armijo condition, used in the MLE step.}

\item{maxit}{Maximum number of iterations for augmentation algorithm. Defaults to 250.}

\item{maxit_fs}{Maximum number of iterations for Fisher scoring. Defaults to 5.}
}
\value{
The optimal beta values under the null or alternative model.
}
\description{
Optimization under null or alternative for multinomial model with a Firth penalty.
}
\author{
Sarah Teichman
}
