% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_test.R
\name{multinom_test}
\alias{multinom_test}
\title{Robust score (Rao) tests for multinomial regression.}
\usage{
multinom_test(
  X = NULL,
  Y,
  formula = NULL,
  data = NULL,
  strong = FALSE,
  j = NULL,
  all_score = FALSE,
  penalty = FALSE,
  pseudo_inv = FALSE
)
}
\arguments{
\item{X}{A \eqn{n x p} design matrix of covariates.}

\item{Y}{A \eqn{n x J} data matrix of outcomes.}

\item{formula}{a one-sided formula specifying the form of the mean model to be fit (use with \code{data} argument if \code{X} is not included)}

\item{data}{a dataframe with \eqn{n} rows containing variables given in \code{formula} (use with \code{formula} argument if \code{X} is not included)}

\item{strong}{If FALSE, this function will compute the robust score statistic to test the weak null that for one specific \eqn{j}, \eqn{\beta_j = 0} for the length \eqn{p} vector \eqn{\beta_j}.
If TRUE, this function instead computes the robust score statistic to test the strong null that \eqn{\beta_1 = \beta_2 = \dots = \beta_{J-1} = 0} for all length \eqn{p} vectors \eqn{\beta_j}, \eqn{j\in\{1,\ldots,J-1\}}.
Default is FALSE.}

\item{j}{If \code{strong} is FALSE, this argument must be supplied. This gives the category \eqn{j} in the weak null hypothesis that \eqn{\beta_j = 0}.}

\item{all_score}{This specifies is scores tests for each individual covariate and category pair should be run. Default is FALSE.}

\item{penalty}{If TRUE will apply a Firth penalty to estimation under the alternative and under the null. Defaults to FALSE (ask Amy her preference)}

\item{pseudo_inv}{Use pseudo inverse for inverted portion of the robust score test to avoid issues with nearly singular matrices.}
}
\value{
The robust score test statistic for the specified hypothesis test. A list including the test statistic, p-value,
estimated parameters under the null hypothesis, and estimated parameters under the alternative hypothesis.
}
\description{
Robust score (Rao) tests for multinomial regression.
}
\author{
Shirley Mathur
}
