% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_robust_score_test.R
\name{null2objective}
\alias{null2objective}
\title{Objective function to find MLE under global null (\beta_1 = beta_2 = \dots = \beta_{J-1} = 0)}
\usage{
null2objective(betanots, Y, X)
}
\arguments{
\item{betanots}{A vector containing the initial values for all \beta_{k0}, for k = 1, \dots, J-1.
In particular, this vector should be so that the entries are ordered as \beta_{10}, \beta_{20}, \dots, \beta_{(J-1)0}.}

\item{Y}{This should be the n x J data matrix of outcomes.}

\item{X}{This should be the n x p design matrix of covariates.}
}
\value{
A vector containing the optimal values for \code{betanots}  to maximize the log-likelihood under the null constraint that \beta_1 = beta_2 = \dots = \beta_{J-1} = 0.
The components are listed out in the same manner as in the \code{betanots} parameter.
}
\description{
Objective function to find MLE under global null (\beta_1 = beta_2 = \dots = \beta_{J-1} = 0)
}
