% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_matrix_contribution.R
\name{S_matrix_contribution}
\alias{S_matrix_contribution}
\title{Compute S matrix contribution for a given cluster to robust score statistic for glm robust score tests.}
\usage{
S_matrix_contribution(indices, model_fits, yy, xx, family, link)
}
\arguments{
\item{indices}{Indices of observations in cluster.}

\item{model_fits}{The fitted glm under the null hypothesis.}

\item{yy}{Vector of observed responses.}

\item{xx}{Design matrix for model.}

\item{family}{The model family for the fitted glm.}

\item{link}{The link function utilized in the fitted glm.}
}
\description{
Compute S matrix contribution for a given cluster to robust score statistic for glm robust score tests.
}
