% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_contribution.R
\name{score_contribution}
\alias{score_contribution}
\title{Compute score contribution to robust score statistic for glm robust score tests.}
\usage{
score_contribution(i, model_fits, yy, xx, family, link)
}
\arguments{
\item{i}{Index of observation.}

\item{model_fits}{The fitted glm under the null hypothesis.}

\item{yy}{Vector of observed responses.}

\item{xx}{Design matrix for model.}

\item{family}{The model family for the fitted glm.}

\item{link}{The link function utilized in the fitted glm.}
}
\description{
Compute score contribution to robust score statistic for glm robust score tests.
}
