% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_lin_com.R
\name{set_up_lin_com}
\alias{set_up_lin_com}
\title{Create matrix to be populated with coefficients for user-specified number of hypotheses.}
\usage{
set_up_lin_com(J, p, n_hypotheses)
}
\arguments{
\item{J}{Number of categories.}

\item{p}{Number of coefficients (excluding intercept).}

\item{n_hypotheses}{Number of hypotheses to test.}
}
\value{
An 'A' matrix with \code{n_hypotheses} rows and \verb{(J-1) x (p+1)} columns with the various coefficient and category combinations as column names.
The matrix is filled with all 0's, but can be modified by the user to reflect the linear combinations they wish to test.
}
\description{
Create matrix to be populated with coefficients for user-specified number of hypotheses.
}
\author{
Shirley Mathur
}
