% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_matrix_contribution.R
\name{V_matrix_contribution}
\alias{V_matrix_contribution}
\title{Compute V matrix contribution to robust score statistic for glm robust score tests.}
\usage{
V_matrix_contribution(
  indices,
  model_fits,
  yy,
  xx,
  pp0 = 1,
  corr_mat,
  family,
  link
)
}
\arguments{
\item{indices}{Indices of observations for cluster.}

\item{model_fits}{The fitted glm under the null hypothesis for the given indices.}

\item{yy}{Vector of observed responses.}

\item{xx}{Design matrix for model.}

\item{pp0}{Number of fixed parameters under null hypothesis.}

\item{corr_mat}{Working correlation matrix for model fit.}

\item{family}{The model family for the fitted glm.}

\item{link}{The link function utilized in the fitted glm.}
}
\description{
Compute V matrix contribution to robust score statistic for glm robust score tests.
}
