% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_test.R
\name{glm_test}
\alias{glm_test}
\title{Generalized Linear Models with robust and non-robust Wald and Rao (score) tests}
\usage{
glm_test(...)
}
\arguments{
\item{...}{Arguments that you would pass to a regular \code{glm} call. Note that for now we only have functionality for Poisson tests with log link}
}
\description{
Generalized Linear Models with robust and non-robust Wald and Rao (score) tests
}
\examples{
# glm_test(dist ~ speed, data = cars, family=poisson(link="log"))


}
