% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.raoFit.R
\name{print.raoFit}
\alias{print.raoFit}
\title{Print function}
\usage{
\method{print}{raoFit}(x)
}
\arguments{
\item{x}{Object of class \code{raoFit}}
}
\value{
\code{NULL}. Displays printed model summary.
}
\description{
Print function
}
\author{
Shirley Mathur
}
