% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_score_test.R
\name{robust_score_test}
\alias{robust_score_test}
\title{Robust score (Rao) tests with finite-sample correction}
\usage{
robust_score_test(glm_object, call_to_model, param = 1, id = NA)
}
\arguments{
\item{glm_object}{The fitted glm under the alternative.}

\item{call_to_model}{The call used to fit the model. Used internally.}

\item{param}{which parameter do you want to test? Used internally.}

\item{id}{observations with the same id are in the same cluster}
}
\description{
The default behavior is to do no finite sample correction (for the covariance of the score) for correlated data,
but to do it for uncorrelated data. This choice performed best under a small simulation study. See Guo et al
for details; the proposed modification is in equation 20.
}
\references{
Guo, X., Pan, W., Connett, J. E., Hannan, P. J., & French, S. A. (2005).
Small-sample performance of the robust score test and its modifications in
generalized estimating equations. \emph{Statistics in Medicine, 24}(22), 3479–3495.
Wiley Online Library. \url{doi:10.1002/sim.2161}
}
