% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_score_test.R
\name{robust_score_test}
\alias{robust_score_test}
\title{Robust score (Rao) test for Poisson regression}
\usage{
robust_score_test(my_formula, df, param = 1)
}
\arguments{
\item{my_formula}{The model under the alternative}

\item{df}{A dataframe or tibble required to fit the model}

\item{param}{which parameter do you want to test?}
}
\description{
Robust score (Rao) test for Poisson regression
}
