% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_lin_com.R
\name{set_up_lin_com}
\alias{set_up_lin_com}
\title{Create matrix to be populated with coefficients for user-specified number of hypotheses.}
\usage{
set_up_lin_com(J, p, n_hypotheses)
}
\arguments{
\item{J}{Number of categories.}

\item{p}{Number of coefficients (excluding intercept).}

\item{n_hypotheses}{Number of hypotheses to test.}
}
\value{
The multinomial probabilities for a given value of \eqn{\beta}.
}
\description{
Create matrix to be populated with coefficients for user-specified number of hypotheses.
}
\author{
Shirley Mathur
}
